+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Refactor `AbstractStandardProcessor.processTypeElements()` to work when the annotation processor is responsible for processing multiple different top-level types.
+
 ### [v0.37](https://github.com/realityforge/proton/tree/v0.37) (2020-02-13) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.36...v0.37)
 
 Changes in this release: